+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+Fix: `Form`: 
+  - Remplace `tiny, regular, default, big` by `small, medium, large` prop size values.
+
 ## [9.4.0] - 2022-03-30
 
 Feature: